Create RoomTempControlSystem_Code
The goal of our project is to create an intelligent device that detects and
displays the current room temperature and accordingly switches on/off the fan or
the heater depending on the desired temperature of the room. It is an Arduino based
control system that uses LCD to display the temperature of the room, Heater (LED)
to increase the temperature and a Fan (Motor) to decrease the temperature.